requirements.txt
Django==2.2.24
environs==5.2.1
dj-database-url==0.5.0
django-phonenumber-field==2.0.
